requirements.txt
flask
gunicorn
jinja2
python-dotenv
python-multipart
azure-identity
azure-keyvault-secrets